refactor: move new concurrent map to its own package
